nest_ssl_project/requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchaudio>=2.0.0
lightning>=2.0.0,<3.0.0
torchmetrics>=0.11.0

# Configuration Management
hydra-core>=1.3.0,<2.0.0
omegaconf>=2.3.0,<3.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.9.0
resampy>=0.4.0

# Data Processing
numpy>=1.22.0
pandas>=1.5.0

# Utilities
tqdm>=4.64.0
tensorboard>=2.10.0
ruamel.yaml>=0.17.0

# Optional but recommended
einops>=0.6.0  # For tensor operations
packaging>=21.0
braceexpand>=0.1.7  # For file path expansion

# Optional: GPU monitoring (for hydra_runner GPU name resolution)
# pynvml>=11.0.0  # Uncomment if you want GPU name resolution in config

# Optional: Advanced dataset processing
# lhotse>=1.31.1  # Uncomment if you want to use Lhotse datasets

# Optional: Experiment tracking
# wandb>=0.15.0  # Uncomment if you want to use Weights & Biases logging

# Development tools (optional)
# pytest>=7.0.0  # For testing
# black>=22.0.0  # For code formatting
# flake8>=5.0.0  # For linting

